vibebot/requirements.txt
av==10.0.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dlib==19.24.1
easydict==1.10
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.12.0
fonttools==4.39.3
gdown==4.7.1
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
llvmlite==0.40.0
MarkupSafe==2.1.2
matplotlib==3.7.1
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.1
numba==0.57.0
numpy==1.24.3
openai-whisper @ git+https://github.com/openai/whisper.git@248b6cb124225dd263bb9bd32d060b6517e067f8
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
psutil==5.9.5
PyAudio==0.2.13
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.22.2
six==1.16.0
soupsieve==2.4.1
SpeechRecognition==3.10.0
sympy==1.11.1
tellopy==0.6.0
thop==0.1.1.post2209072238
tiktoken==0.3.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
ultralytics==8.0.96
urllib3==1.26.15
